+++ TestsDelivery.DAL/Exceptions/Subjects/SubjectNotFoundException.cs
@@ -4,7 +4,7 @@ namespace TestsDelivery.DAL.Exceptions.Subjects
 {
     public class SubjectNotFoundException : Exception
     {
-        public SubjectNotFoundException(int id)
+        public SubjectNotFoundException(long id)
             : base($"Subject with id = {id} does not exists.")
         {
         }
