+++ TestsDelivery.DAL/Models/Questions/AnswerOption.cs
@@ -6,14 +6,14 @@ namespace TestsDelivery.DAL.Models.Questions
     public class AnswerOption
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Text { get; set; }
 
         public bool IsCorrect { get; set; }
 
         [ForeignKey(nameof(Questions.Question))]
-        public int QuestionId { get; set; }
+        public long QuestionId { get; set; }
 
         public virtual Question Question { get; set; }
     }
