+++ TestsDelivery.DAL/Repositories/Subjects/ISubjectsRepository.cs
@@ -6,7 +6,7 @@ namespace TestsDelivery.DAL.Repositories.Subjects
     {
         void CreateSubject(Subject subject);
 
-        Subject GetSubject(int id);
+        Subject GetSubject(long id);
 
         void EditSubject(Subject subject);
     }
